requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-core>=0.1.21
langchain-chroma>=0.1.0
openai>=1.12.0
chromadb>=0.4.22
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.4.0
scikit-learn>=1.3.0
networkx>=3.0
numpy>=1.24.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
boto3>=1.34.0 